backend/requirements.txt
asgiref==3.7.2
boto3==1.28.9
botocore==1.31.9
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
decorator==5.1.1
dj-database-url==2.0.0
Django==4.2.3
django-cors-headers==4.2.0
django-dotenv==1.4.2
django-environ==0.10.0
django-filter==23.2
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
future==0.18.3
geocoder==1.38.1
gunicorn==21.2.0
idna==3.4
jmespath==1.0.1
packaging==23.1
psycopg2==2.9.6
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3
ratelim==0.1.6
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==1.26.16
whitenoise==6.5.0
